refactor: admin package 정리
