+++ admin/src/main/kotlin/com/daegusw/apply/admin/controller/AdminFirstGradeController.kt
@@ -1,7 +1,6 @@
-package com.daegusw.apply.admin.domain.admin.controller
+package com.daegusw.apply.admin.controller
 
-import com.daegusw.apply.admin.domain.admin.dto.FirstTableResponse
-import com.daegusw.apply.admin.domain.admin.service.AdminFirstGradeService
+import com.daegusw.apply.admin.dto.FirstTableResponse
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
@@ -9,7 +8,7 @@ import org.springframework.web.bind.annotation.RestController
 @RestController
 @RequestMapping("/admin/first")
 class AdminFirstGradeController(
-    private val adminFirstGradeService: AdminFirstGradeService
+    private val adminFirstGradeService: com.daegusw.apply.admin.service.AdminFirstGradeService
 ) {
 
     @GetMapping
