+++ admin/src/main/kotlin/com/daegusw/apply/admin/controller/AdminPassApplicantController.kt
@@ -1,7 +1,7 @@
-package com.daegusw.apply.admin.domain.admin.controller
+package com.daegusw.apply.admin.controller
 
-import com.daegusw.apply.admin.domain.admin.dto.PassApplicantTableResponse
-import com.daegusw.apply.admin.domain.admin.service.AdminPassApplicantService
+import com.daegusw.apply.admin.dto.PassApplicantTableResponse
+import com.daegusw.apply.admin.service.AdminPassApplicantService
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
