+++ admin/src/main/kotlin/com/daegusw/apply/admin/controller/AdminStatisticsController.kt
@@ -1,10 +1,9 @@
-package com.daegusw.apply.admin.domain.admin.controller
+package com.daegusw.apply.admin.controller
 
-import com.daegusw.apply.admin.domain.admin.dto.ApplicantCompetitionRateResponse
-import com.daegusw.apply.admin.domain.admin.dto.ApplicantGraduationTypeCompetitionRateResponse
-import com.daegusw.apply.admin.domain.admin.dto.ApplicantSchoolCompetitionRateResponse
-import com.daegusw.apply.admin.domain.admin.dto.UserSchoolCityInfoResponse
-import com.daegusw.apply.admin.domain.admin.service.AdminStatisticsService
+import com.daegusw.apply.admin.dto.ApplicantGraduationTypeCompetitionRateResponse
+import com.daegusw.apply.admin.dto.ApplicantSchoolCompetitionRateResponse
+import com.daegusw.apply.admin.dto.UserSchoolCityInfoResponse
+import com.daegusw.apply.admin.service.AdminStatisticsService
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
@@ -16,7 +15,7 @@ class AdminStatisticsController(
 ) {
 
     @GetMapping("/applicant")
-    fun applyCompetitionRate(): List<ApplicantCompetitionRateResponse> {
+    fun applyCompetitionRate(): List<com.daegusw.apply.admin.dto.ApplicantCompetitionRateResponse> {
         return adminStatisticsService.getApplicantCompetitionRate()
     }
