+++ admin/src/main/kotlin/com/daegusw/apply/admin/dto/ApplicantSchoolCompetitionRateResponse.kt
@@ -1,4 +1,4 @@
-package com.daegusw.apply.admin.domain.admin.dto
+package com.daegusw.apply.admin.dto
 
 data class ApplicantSchoolCompetitionRateResponse(
     val apply: String,
