+++ admin/src/main/kotlin/com/daegusw/apply/admin/dto/PassApplicantTableResponse.kt
@@ -1,4 +1,4 @@
-package com.daegusw.apply.admin.domain.admin.dto
+package com.daegusw.apply.admin.dto
 
 import com.daegusw.apply.admission.domain.admission.Admission
 import com.daegusw.apply.admission.domain.admission.constant.ApplyType
