+++ admin/src/main/kotlin/com/daegusw/apply/admin/dto/excel/AptitudeExcel.kt
@@ -1,6 +1,5 @@
-package com.daegusw.apply.admin.domain.admin.dto.excel
+package com.daegusw.apply.admin.dto.excel
 
-import com.daegusw.apply.admin.domain.secondary.domain.Secondary
 import com.daegusw.apply.applicant.domain.applicant.Applicant
 import com.daegusw.apply.core.excel.common.annotation.DefaultBodyStyle
 import com.daegusw.apply.core.excel.common.annotation.DefaultHeaderStyle
@@ -8,10 +7,11 @@ import com.daegusw.apply.core.excel.common.annotation.ExcelColumn
 import com.daegusw.apply.core.excel.common.annotation.ExcelColumnStyle
 import com.daegusw.apply.core.excel.style.DefaultBodyCellStyle
 import com.daegusw.apply.core.excel.style.DefaultHeaderCellStyle
+import com.daegusw.apply.secondary.domain.Secondary
 
 @DefaultHeaderStyle(style = ExcelColumnStyle(excelCellStyleClass = DefaultHeaderCellStyle::class))
 @DefaultBodyStyle(style = ExcelColumnStyle(excelCellStyleClass = DefaultBodyCellStyle::class))
-data class InterviewExcel(
+data class AptitudeExcel(
     @ExcelColumn(headerName = "순번", width = 2000)
     val index: Int,
     @ExcelColumn(headerName = "접수 번호", width = 3000)
@@ -20,16 +20,13 @@ data class InterviewExcel(
     val examNumber: Int? = 0,
     @ExcelColumn(headerName = "성명", width = 2000)
     val name: String? = "",
-    @ExcelColumn(headerName = "학업 및 진로 역량", width = 4000)
-    val studyCompetence: Double? = 0.0,
-    @ExcelColumn(headerName = "컴퓨팅 사고력 역량", width = 4000)
-    val computingCompetence: Double? = 0.0
+    @ExcelColumn(headerName = "직무적성 소양평가", width = 4000)
+    val score: Double? = 0.0
 ) {
 
-    constructor(index: Int, applicant: Applicant?, secondary: Secondary?) : this (
+    constructor(index: Int, applicant: Applicant?, secondary: Secondary?) : this(
         index = index,
         name = applicant?.privacy?.name,
-        studyCompetence = secondary?.interview?.studyCompetence,
-        computingCompetence = secondary?.interview?.computingCompetence
+        score = secondary?.aptitude?.jobScore
     )
 }
