+++ admin/src/main/kotlin/com/daegusw/apply/admin/service/AdminSecondaryGradeService.kt
@@ -1,9 +1,8 @@
-package com.daegusw.apply.admin.domain.admin.service
+package com.daegusw.apply.admin.service
 
-import com.daegusw.apply.admin.domain.admin.dto.SecondaryTableResponse
-import com.daegusw.apply.admin.domain.admin.dto.excel.AptitudeExcel
-import com.daegusw.apply.admin.domain.admin.dto.excel.InterviewExcel
-import com.daegusw.apply.admin.domain.secondary.repository.SecondaryRepository
+import com.daegusw.apply.admin.dto.SecondaryTableResponse
+import com.daegusw.apply.admin.dto.excel.AptitudeExcel
+import com.daegusw.apply.admin.dto.excel.InterviewExcel
 import com.daegusw.apply.admission.application.port.out.persistence.QueryAdmissionPort
 import com.daegusw.apply.admission.domain.admission.constant.Progress
 import com.daegusw.apply.applicant.application.port.out.persistence.QueryApplicantPort
@@ -13,19 +12,21 @@ import org.springframework.stereotype.Service
 class AdminSecondaryGradeService(
     private val queryApplicantPort: QueryApplicantPort,
     private val queryAdmissionPort: QueryAdmissionPort,
-    private val secondarySpi: SecondaryRepository
+//    private val secondarySpi: SecondaryRepository
 ) {
 
     fun getSecondaryTable(): List<SecondaryTableResponse> {
         val tableResponse: MutableList<SecondaryTableResponse> = mutableListOf()
-        val admission = queryAdmissionPort.findAllByProgress(Progress.FIRST)
+        val admission = queryAdmissionPort.findAllByProgress(Progress.FIRST_PASSED)
         var index = 1
         admission.forEach {
             val applicant = queryApplicantPort.findById(it.applicant?.member!!)
-            val secondary = secondarySpi.findByMemberId(it.applicant?.member!!)
+            //TODO Secondary Repository
+//           val secondary = secondarySpi.findById(it.applicant?.member!!)
+
             tableResponse.add(
                 SecondaryTableResponse(
-                    index++, applicant, it, secondary
+                    index++, applicant, it, TODO()
                 )
             )
         }
@@ -35,13 +36,13 @@ class AdminSecondaryGradeService(
 
     fun jobExcelDownload(): List<AptitudeExcel> {
         val response: MutableList<AptitudeExcel> = mutableListOf()
-        val admission = queryAdmissionPort.findAllByProgress(Progress.FIRST)
+        val admission = queryAdmissionPort.findAllByProgress(Progress.FIRST_PASSED)
         var index = 1
         admission.forEach {
             val applicant = queryApplicantPort.findById(it.applicant?.member!!)
-            val secondary = secondarySpi.findByMemberId(it.applicant?.member!!)
+/*            val secondary = secondarySpi.findByMemberId(it.applicant?.member!!)*/
             response.add(
-                AptitudeExcel(index++, applicant, secondary)
+                AptitudeExcel(index++, applicant, TODO("secondary"))
             )
         }
 
@@ -50,13 +51,13 @@ class AdminSecondaryGradeService(
 
     fun interviewExcelDownload(): List<InterviewExcel> {
         val response: MutableList<InterviewExcel> = mutableListOf()
-        val admission = queryAdmissionPort.findAllByProgress(Progress.FIRST)
+        val admission = queryAdmissionPort.findAllByProgress(Progress.FIRST_PASSED)
         var index = 1
         admission.forEach {
             val applicant = queryApplicantPort.findById(it.applicant?.member!!)
-            val secondary = secondarySpi.findByMemberId(it.applicant?.member!!)
+            /*val secondary = secondarySpi.findByMemberId(it.applicant?.member!!)*/
             response.add(
-                InterviewExcel(index++, applicant, secondary)
+                InterviewExcel(index++, applicant, TODO("secondary"))
             )
         }
